+++ manage.py
@@ -2,7 +2,7 @@
 import os
 import sys
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "test_project.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.dev_settings")
 
 if __name__ == "__main__":
     try: